% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geomod.R
\docType{data}
\name{data}
\alias{data}
\title{Step3: Import the data and explore data structures}
\format{An object of class \code{SpatialPointsDataFrame} with 42 rows and 6 columns.}
\usage{
data
}
\description{
Step3: Import the data and explore data structures
}
\keyword{datasets}
