% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geomod.R
\docType{data}
\name{por.lme}
\alias{por.lme}
\title{Step 7 - Soil Property Prediction Model (Regression)}
\format{An object of class \code{lm} of length 12.}
\usage{
por.lme
}
\description{
Step 7 - Soil Property Prediction Model (Regression)
}
\keyword{datasets}
