% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geomod.R
\docType{data}
\name{por.rev}
\alias{por.rev}
\title{Step 8 - Variogram analysis of the autocorrelation}
\format{An object of class \code{gstatVariogram} (inherits from \code{data.frame}) with 14 rows and 6 columns.}
\usage{
por.rev
}
\description{
Step 8 - Variogram analysis of the autocorrelation
}
\keyword{datasets}
