% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geomod.R
\docType{data}
\name{predictors.ov}
\alias{predictors.ov}
\title{Step 6 - Overlay the environmental correlates}
\format{An object of class \code{data.frame} with 42 rows and 3 columns.}
\usage{
predictors.ov
}
\description{
Step 6 - Overlay the environmental correlates
}
\keyword{datasets}
