% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geomod.R
\docType{data}
\name{predictors}
\alias{predictors}
\title{Step 5: Import the environmental correlates and convert them into spatial dataframe}
\format{An object of class \code{SpatialGridDataFrame} with 84148 rows and 4 columns.}
\usage{
predictors
}
\description{
Step 5: Import the environmental correlates and convert them into spatial dataframe
}
\keyword{datasets}
